requirements.txt
paddlepaddle
sklearn
pandas
numpy
matplotlib